chore: Refactor DdMapingFileUploadTask
Remove duplication and have DdMappingFileUploadTask inherit from FileUploadTask
