+++ dd-sdk-android-gradle-plugin/src/main/kotlin/com/datadog/gradle/plugin/DdMappingFileUploadTask.kt
@@ -7,23 +7,9 @@
 package com.datadog.gradle.plugin
 
 import com.datadog.gradle.plugin.DdAndroidGradlePlugin.Companion.LOGGER
-import com.datadog.gradle.plugin.internal.ApiKeySource
-import com.datadog.gradle.plugin.internal.DdAppIdentifier
-import com.datadog.gradle.plugin.internal.OkHttpUploader
 import com.datadog.gradle.plugin.internal.Uploader
-import org.gradle.api.DefaultTask
-import org.gradle.api.provider.Provider
 import org.gradle.api.provider.ProviderFactory
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputFile
-import org.gradle.api.tasks.InputFiles
-import org.gradle.api.tasks.Internal
-import org.gradle.api.tasks.Optional
-import org.gradle.api.tasks.OutputFile
-import org.gradle.api.tasks.TaskAction
-import org.json.JSONArray
-import org.json.JSONException
-import org.json.JSONObject
 import java.io.File
 import javax.inject.Inject
 
@@ -33,69 +19,8 @@ import javax.inject.Inject
 open class DdMappingFileUploadTask
 @Inject constructor(
     providerFactory: ProviderFactory,
-    @get:Internal internal val repositoryDetector: RepositoryDetector
-) : DefaultTask() {
-
-    @get:Internal
-    internal var uploader: Uploader = OkHttpUploader()
-
-    // TODO RUMM-2382 use Provider<String>
-    /**
-     * The API Key used to upload the data.
-     */
-    @get:Input
-    var apiKey: String = ""
-
-    private val disableGzipOption: Provider<String> =
-        providerFactory.gradleProperty(DISABLE_GZIP_GRADLE_PROPERTY)
-
-    /**
-     * Source of the API key set: environment, gradle property, etc.
-     */
-    @get:Input
-    var apiKeySource: ApiKeySource = ApiKeySource.NONE
-
-    /**
-     * The variant name of the application.
-     */
-    @get:Input
-    var variantName: String = ""
-
-    /**
-     * The version name of the application.
-     */
-    @get:Input
-    var versionName: String = ""
-
-    /**
-     * The version code of the application.
-     */
-    @get:Input
-    var versionCode: Int = 0
-
-    /**
-     * The service name of the application (by default, it is your app's package name).
-     */
-    @get:Input
-    var serviceName: String = ""
-
-    /**
-     * The Datadog site to upload to (one of "US1", "EU1", "US1_FED").
-     */
-    @get:Input
-    var site: String = ""
-
-    /**
-     * The url of the remote repository where the source code was deployed.
-     */
-    @get:Input
-    var remoteRepositoryUrl: String = ""
-
-    /**
-     * Build ID which will be used for mapping file matching.
-     */
-    @get:Input
-    var buildId: Provider<String> = providerFactory.provider { "" }
+    repositoryDetector: RepositoryDetector
+) : DdFileUploadTask(providerFactory, repositoryDetector) {
 
     /**
      * The path to the mapping file to upload.
@@ -115,25 +40,6 @@ open class DdMappingFileUploadTask
     @get:Input
     var mappingFileTrimIndents: Boolean = false
 
-    /**
-     * datadog-ci.json file, if found or applicable for the particular task.
-     */
-    @Optional
-    @get:InputFile
-    var datadogCiFile: File? = null
-
-    /**
-     * The sourceSet root folders.
-     */
-    @get:InputFiles
-    var sourceSetRoots: List<File> = emptyList()
-
-    /**
-     * The file containing the repository description.
-     */
-    @get:OutputFile
-    var repositoryFile: File = File("")
-
     init {
         group = DdAndroidGradlePlugin.DATADOG_TASK_GROUP
         description = "Uploads the Proguard/R8 mapping file to Datadog"
@@ -141,162 +47,21 @@ open class DdMappingFileUploadTask
         outputs.upToDateWhen { false }
     }
 
-    // region Task
-
-    /**
-     * Uploads the mapping file to Datadog.
-     */
-    @TaskAction
-    fun applyTask() {
-        datadogCiFile?.let {
-            applyDatadogCiConfig(it)
-        }
-        applySiteFromEnvironment()
-        validateConfiguration()
-
-        check(!(apiKey.contains("\"") || apiKey.contains("'"))) {
-            INVALID_API_KEY_FORMAT_ERROR
-        }
-
-        check(buildId.isPresent && buildId.get().isNotEmpty()) {
-            MISSING_BUILD_ID_ERROR
-        }
-
+    override fun getFilesList(): List<Uploader.UploadFileInfo> {
         var mappingFile = File(mappingFilePath)
-        if (!validateMappingFile(mappingFile)) return
+        if (!validateMappingFile(mappingFile)) return emptyList()
 
         mappingFile = shrinkMappingFile(mappingFile)
 
-        val repositories = repositoryDetector.detectRepositories(
-            sourceSetRoots,
-            remoteRepositoryUrl
-        )
-        if (repositories.isNotEmpty()) {
-            generateRepositoryFile(repositories)
-        }
-
-        val site = DatadogSite.valueOf(site)
-        val fileInfo = Uploader.UploadFileInfo(
-            fileKey = KEY_JVM_MAPPING_FILE,
-            file = mappingFile,
-            encoding = MEDIA_TYPE_TXT,
-            fileType = TYPE_JVM_MAPPING_FILE,
-            fileName = KEY_JVM_MAPPING_FILE_NAME
-        )
-        uploader.upload(
-            site,
-            fileInfo,
-            if (repositories.isEmpty()) null else repositoryFile,
-            apiKey,
-            DdAppIdentifier(
-                serviceName = serviceName,
-                version = versionName,
-                versionCode = versionCode,
-                variant = variantName,
-                buildId = buildId.get()
-            ),
-            repositories.firstOrNull(),
-            !disableGzipOption.isPresent
-        )
-    }
-
-    private fun applySiteFromEnvironment() {
-        val environmentSite = System.getenv(DATADOG_SITE)
-        if (!environmentSite.isNullOrEmpty()) {
-            if (this.site.isNotEmpty()) {
-                LOGGER.info(
-                    "Site property found as DATADOG_SITE env variable, but it will be ignored," +
-                        " because also an explicit one was provided in extension."
-                )
-                return
-            }
-            val site = DatadogSite.fromDomain(environmentSite)
-            if (site == null) {
-                LOGGER.warn("Unknown Datadog domain provided: $environmentSite, ignoring it.")
-            } else {
-                LOGGER.info("Site property found in Datadog CI config file, using it.")
-                this.site = site.name
-            }
-        }
-    }
-
-    private fun applyDatadogCiConfig(datadogCiFile: File) {
-        try {
-            val config = JSONObject(datadogCiFile.readText())
-            applyApiKeyFromDatadogCiConfig(config)
-            applySiteFromDatadogCiConfig(config)
-        } catch (e: JSONException) {
-            LOGGER.error("Failed to parse Datadog CI config file.", e)
-        }
-    }
-
-    private fun applyApiKeyFromDatadogCiConfig(config: JSONObject) {
-        val apiKey = config.optString(DATADOG_CI_API_KEY_PROPERTY, null)
-        if (!apiKey.isNullOrEmpty()) {
-            if (this.apiKeySource == ApiKeySource.GRADLE_PROPERTY) {
-                LOGGER.info(
-                    "API key found in Datadog CI config file, but it will be ignored," +
-                        " because also an explicit one was provided as a gradle property."
-                )
-            } else {
-                LOGGER.info("API key found in Datadog CI config file, using it.")
-                this.apiKey = apiKey
-                this.apiKeySource = ApiKeySource.DATADOG_CI_CONFIG_FILE
-            }
-        }
-    }
-
-    private fun applySiteFromDatadogCiConfig(config: JSONObject) {
-        if (this.site.isNotEmpty()) {
-            LOGGER.info(
-                "Site property found in Datadog CI config file, but it will be ignored," +
-                    " because also an explicit one was provided in extension."
+        return listOf(
+            Uploader.UploadFileInfo(
+                fileKey = KEY_JVM_MAPPING_FILE,
+                file = mappingFile,
+                encoding = MEDIA_TYPE_TXT,
+                fileType = TYPE_JVM_MAPPING_FILE,
+                fileName = KEY_JVM_MAPPING_FILE_NAME
             )
-            return
-        }
-        val siteAsDomain = config.optString(DATADOG_CI_SITE_PROPERTY, null)
-        if (!siteAsDomain.isNullOrEmpty()) {
-            val site = DatadogSite.fromDomain(siteAsDomain)
-            if (site == null) {
-                LOGGER.warn("Unknown Datadog domain provided: $siteAsDomain, ignoring it.")
-            } else {
-                LOGGER.info("Site property found in Datadog CI config file, using it.")
-                this.site = site.name
-            }
-        }
-    }
-
-    // endregion
-
-    // region Internal
-
-    @Suppress("CheckInternal")
-    private fun validateConfiguration() {
-        check(apiKey.isNotBlank()) { API_KEY_MISSING_ERROR }
-
-        if (site.isBlank()) {
-            site = DatadogSite.US1.name
-        } else {
-            val validSiteIds = DatadogSite.validIds
-            check(site in validSiteIds) {
-                "You need to provide a valid site (one of ${validSiteIds.joinToString()})"
-            }
-        }
-    }
-
-    private fun generateRepositoryFile(repositories: List<RepositoryInfo>) {
-        val data = JSONArray()
-        repositories.forEach {
-            data.put(it.toJson())
-            LOGGER.info("Detected repository:\n${it.toJson().toString(INDENT)}")
-        }
-
-        val jsonObject = JSONObject()
-        jsonObject.put("version", 1)
-        jsonObject.put("data", data)
-
-        repositoryFile.parentFile.mkdirs()
-        repositoryFile.writeText(jsonObject.toString(0))
+        )
     }
 
     @Suppress("CheckInternal")
@@ -386,8 +151,6 @@ open class DdMappingFileUploadTask
     // endregion
 
     internal companion object {
-        private const val INDENT = 4
-
         internal const val TYPE_JVM_MAPPING_FILE = "jvm_mapping_file"
         internal const val KEY_JVM_MAPPING_FILE = "jvm_mapping_file"
         internal const val KEY_JVM_MAPPING_FILE_NAME = "jvm_mapping"
@@ -396,11 +159,6 @@ open class DdMappingFileUploadTask
         private const val MAPPING_FILE_CHANGE_DELIMITER = "->"
         private const val MAPPING_FILE_COMMENT_CHAR = '#'
         const val MAPPING_OPTIMIZED_FILE_NAME = "mapping-shrinked.txt"
-
-        private const val DATADOG_CI_API_KEY_PROPERTY = "apiKey"
-        private const val DATADOG_CI_SITE_PROPERTY = "datadogSite"
-        const val DATADOG_SITE = "DATADOG_SITE"
-        const val DISABLE_GZIP_GRADLE_PROPERTY = "dd-disable-gzip"
 
         const val API_KEY_MISSING_ERROR = "Make sure you define an API KEY to upload your mapping files to Datadog. " +
             "Create a DD_API_KEY or DATADOG_API_KEY environment variable, gradle" +
