+++ dd-sdk-android-gradle-plugin/src/main/kotlin/com/datadog/gradle/plugin/DdNdkSymbolFileUploadTask.kt
@@ -2,7 +2,6 @@ package com.datadog.gradle.plugin
 
 import com.android.build.gradle.api.ApplicationVariant
 import com.android.build.gradle.tasks.ExternalNativeBuildTask
-import com.android.builder.model.Version
 import com.datadog.gradle.plugin.internal.ApiKey
 import com.datadog.gradle.plugin.internal.Uploader
 import org.gradle.api.Project
@@ -77,52 +76,17 @@ internal abstract class DdNdkSymbolFileUploadTask @Inject constructor(
         internal const val TYPE_NDK_SYMBOL_FILE = "ndk_symbol_file"
         internal const val ENCODING = "application/octet-stream"
         internal val SUPPORTED_ARCHS = setOf("armeabi-v7a", "arm64-v8a", "x86", "x86_64")
-        internal const val MAX_DATADOG_CI_FILE_LOOKUP_LEVELS = 4
 
         internal val LOGGER = Logging.getLogger("DdSymbolFileUploadTask")
-
-        @Suppress("StringLiteralDuplication")
-        private fun resolveDatadogRepositoryFile(target: Project): File {
-            val outputsDir = File(target.buildDir, "outputs")
-            val reportsDir = File(outputsDir, "reports")
-            val datadogDir = File(reportsDir, "datadog")
-            return File(datadogDir, "repository.json")
-        }
-
-        internal fun findDatadogCiFile(projectDir: File): File? {
-            var currentDir: File? = projectDir
-            var levelsUp = 0
-            while (currentDir != null && levelsUp < MAX_DATADOG_CI_FILE_LOOKUP_LEVELS) {
-                val datadogCiFile = File(currentDir, "datadog-ci.json")
-                if (datadogCiFile.exists()) {
-                    return datadogCiFile
-                }
-                currentDir = currentDir.parentFile
-                levelsUp++
-            }
-            return null
-        }
-
-        @Suppress("MagicNumber", "ReturnCount")
-        private fun isAgpAbove(major: Int, minor: Int, patch: Int): Boolean {
-            val version = Version.ANDROID_GRADLE_PLUGIN_VERSION
-            val groups = version.split(".")
-            if (groups.size < 3) return false
-            val currentMajor = groups[0].toIntOrNull()
-            val currentMinor = groups[1].toIntOrNull()
-            val currentPatch = groups[2].substringBefore("-").toIntOrNull()
-            if (currentMajor == null || currentMinor == null || currentPatch == null) return false
-            return currentMajor >= major && currentMinor >= minor && currentPatch >= patch
-        }
 
         private fun getSearchDirs(
             buildTask: TaskProvider<ExternalNativeBuildTask>,
             providerFactory: ProviderFactory
         ): Provider<File?> {
             return buildTask.flatMap { task ->
-                // var soFolder: Provider
+                // var soFolder: `Provider
                 @Suppress("MagicNumber")
-                if (isAgpAbove(8, 0, 0)) {
+                if (DdTaskUtils.isAgpAbove(8, 0, 0)) {
                     task.soFolder.map { it.asFile }
                 } else {
                     val soFolder = ExternalNativeBuildTask::class.memberProperties.find {
@@ -163,7 +127,7 @@ internal abstract class DdNdkSymbolFileUploadTask @Inject constructor(
                     variant.sourceSets.forEach {
                         roots.addAll(it.javaDirectories)
                         @Suppress("MagicNumber")
-                        if (isAgpAbove(7, 0, 0)) {
+                        if (DdTaskUtils.isAgpAbove(7, 0, 0)) {
                             roots.addAll(it.kotlinDirectories)
                         }
                     }
@@ -176,8 +140,8 @@ internal abstract class DdNdkSymbolFileUploadTask @Inject constructor(
                         task.dependsOn(buildTask)
                     }
 
-                    task.datadogCiFile = findDatadogCiFile(project.rootDir)
-                    task.repositoryFile = resolveDatadogRepositoryFile(project)
+                    task.datadogCiFile = DdTaskUtils.findDatadogCiFile(project.rootDir)
+                    task.repositoryFile = DdTaskUtils.resolveDatadogRepositoryFile(project)
                     task.configureWith(
                         apiKey,
                         extensionConfiguration,
