+++ dd-sdk-android-gradle-plugin/src/test/kotlin/com/datadog/gradle/plugin/DdMappingFileUploadTaskTest.kt
@@ -126,12 +126,12 @@ internal class DdMappingFileUploadTaskTest {
             whenever(isPresent) doReturn true
             whenever(get()) doReturn fakeBuildId
         }
-        setEnv(DdMappingFileUploadTask.DATADOG_SITE, "")
+        setEnv(DdFileUploadTask.DATADOG_SITE, "")
     }
 
     @AfterEach
     fun `tear down`() {
-        removeEnv(DdMappingFileUploadTask.DATADOG_SITE)
+        removeEnv(DdFileUploadTask.DATADOG_SITE)
     }
 
     @Test
@@ -750,7 +750,7 @@ internal class DdMappingFileUploadTaskTest {
     fun `𝕄 read site from environment variable 𝕎 applyTask() {site is not set}`(forge: Forge) {
         // Given
         val fakeDatadogEnvDomain = forge.aValueFrom(DatadogSite::class.java).domain
-        setEnv(DdMappingFileUploadTask.DATADOG_SITE, fakeDatadogEnvDomain)
+        setEnv(DdFileUploadTask.DATADOG_SITE, fakeDatadogEnvDomain)
         testedTask.site = ""
 
         // When
